+++ CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [unreleased]
+# [Unreleased]
 
 ### Fixed
 
 - Fixed a bug introduced in the last release where placemarks wouldn't update
   after changing floors
+- Fixed a bug introduced in the last release where the `loadPlacemarks` toggle
+  didn't work properly
 
 # [0.8.1] - 2020-09-09