+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Unpublished floors are now hidden from the floor selector
 - A bug where placemarks wouldn't update after changing floors
+- A bug where `loadPlacemarks` didn't work correctly
 
 
 ## [0.8.1] - 2020-09-09