+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Placemarks wouldn't update after changing floors
 - `loadPlacemarks` didn't work correctly
 
+### Added
+
+- Several new methods starting with `fetch` on the `MeridianSDK.API` class
+
+### Deprecated
+
+- The `.axios` property on a `MeridianSDK.API` instance
+
 ## [0.8.1] - 2020-09-09
 
 ### Fixed