+++ CHANGELOG.md
@@ -21,6 +21,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `api.fetchFloorsByLocation`
 - `api.fetchSVG`
 
+### Changed
+
+- `api.token` is now read-only
+- `api.environment` is now read-only
+
 ### Deprecated
 
 - `api.axios`