+++ CHANGELOG.md
@@ -13,6 +13,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Placemarks wouldn't update after changing floors
 - `loadPlacemarks` didn't work correctly
 
+### Added
+
+- `api.fetchTagsByFloor`
+- `api.fetchTagsByLocation`
+- `api.fetchPlacemarksByFloor`
+- `api.fetchFloorsByLocation`
+- `api.fetchSVG`
+
+### Changed
+
+- `api.token` is now read-only
+- `api.environment` is now read-only
+
+### Deprecated
+
+- `api.axios`
+
 ## [0.8.1] - 2020-09-09
 
 ### Fixed